src/requirements/local.txt
Django>=3.0.8
djangorestframework==3.11.2
psycopg2-binary
gunicorn==20.0.4
django-cors-headers>=3.5.0
whitenoise==5.2.0
django-storages==1.10
pillow==8.1.2
requests==2.24.0
